Added mock operator and listening from publisher
