+++ publisher.py
@@ -1,84 +1,7 @@
 from config import SOCKET_HOST, SOCKET_PORT
 from icd_config import Command, int_to_bytes
-import socket
+from connections import OperatorConnection
 import time
-import threading
-
-
-class Connection:
-    def __init__(self, host, port):
-        self.host = host
-        self.port = port
-        self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.is_connected = False
-        self.async_connect()
-
-    def async_connect(self):
-        thread = threading.Thread(target=self.connect)
-        thread.start()
-
-    def connect(self):
-        raise NotImplementedError("connect method is not implemented")
-
-    def close_socket(self):
-        self.socket.close()
-        self.is_connected = False
-
-    def publish(self, command: int, payload: bytes = None):
-        """
-        Command ID      UINT32	Unique ID for individual commands
-        self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        RESERVED     	UINT16	RESERVED
-        Command Value	UINT16	Command for device to carry out
-        Length	        UINT16	Length of Payload
-        Payload	        UINT8[]	Command Info
-        CRC	            UINT16	Checksum
-        """
-        # Get a unique, incrementing command id. Increment by 2, so that the response
-        # from the operator always returns odd command ids and the publisher always sends
-        # even command ids. Commands can be associated with each other by checking if they
-        # have the same modulus of 2.
-        command_id = Publisher.command_count
-        Publisher.command_count += 2
-
-        if payload != None:
-            payload_length = len(payload)
-        else:
-            payload_length = 0
-
-        # TODO: Implement checksum. May get removed
-        crc = int_to_bytes(0, num_bits=16, unsigned=True)
-
-        command_id = int_to_bytes(command_id, num_bits=32, unsigned=True)
-        reserved = int_to_bytes(0, num_bits=16, unsigned=True)
-        command = int_to_bytes(command, num_bits=16, unsigned=True)
-        payload_length = int_to_bytes(payload_length, num_bits=16, unsigned=True)
-
-        # Put header together
-        header = command_id + reserved + command + payload_length
-
-        # Put everything together 
-        body = header
-
-        if payload != None:
-            body += payload
-
-        body += crc
-
-        self.socket.send(body)
-
-
-class OperatorConnection(Connection):
-    def connect(self):
-        while(True):
-            try:
-                self.socket.connect((self.host, self.port))
-                self.is_connected = True
-                print("Connected to socket: " + self.host + ":" + str(self.port))
-                break
-            except ConnectionRefusedError: # catch refused connection for retry
-                print("Connection to " + self.host + ":" + str(self.port) + " failed, retrying in 5s")
-                time.sleep(5)
 
 
 class Publisher:
@@ -171,3 +94,18 @@ class Publisher:
             command=int(Command.HOME),
             payload=delay
         )
+
+
+def main():
+    while (not Publisher.connection.is_connected):
+        time.sleep(1)
+
+    Publisher.home(0)
+
+    # Stay alive
+    while True:
+        time.sleep(1)
+
+
+if __name__ == "__main__":
+    main()
